+++ lib/polyfill/v2_5.rb
@@ -1,3 +1,5 @@
+require_relative 'v2_5/integer'
+
 module Polyfill
   module V2_5
   end